+++ .phan/config.php
@@ -605,14 +605,15 @@ return [
         // These are specific to Phan's coding style
         'StrictComparisonPlugin',
         // Warn about `$var == SOME_INT_OR_STRING_CONST` due to unintuitive behavior such as `0 == 'a'`
-        '.phan/plugins/StrictLiteralComparisonPlugin.php',
-        '.phan/plugins/ShortArrayPlugin.php',
-        '.phan/plugins/SimplifyExpressionPlugin.php',
+        'StrictLiteralComparisonPlugin',
+        'ShortArrayPlugin',
+        'SimplifyExpressionPlugin',
         // 'UnknownClassElementAccessPlugin' is more useful with batch analysis than in an editor.
         // It's used in tests/run_test __FakeSelfFallbackTest
 
         // This checks that there are no accidental echos/printfs left inside Phan's code.
-        '.phan/plugins/RemoveDebugStatementPlugin.php',
+        'RemoveDebugStatementPlugin',
+        '.phan/plugins/UnsafeCodePlugin.php',
 
         ////////////////////////////////////////////////////////////////////////
         // End plugins for Phan's self-analysis