+++ tests/plugin_test/.phan/config.php
@@ -148,7 +148,8 @@ return [
         'UnknownClassElementAccessPlugin',
         'DuplicateConstantPlugin',
         'ConstantVariablePlugin',
-        '../../.phan/plugins/ShortArrayPlugin.php',
-        '../../.phan/plugins/SimplifyExpressionPlugin.php',
+        'ShortArrayPlugin',
+        'SimplifyExpressionPlugin',
+        'UnsafeCodePlugin',
     ],
 ];