+++ tests/plugin_test/src/184_shell_exec.php
@@ -1,0 +1,5 @@
+<?php
+if (false) {
+    $hello = 'Hello';
+    $x = `echo $hello,` . `echo "world"`;
+}